database/requirements.txt
# Database dependencies
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
